Code review/feature comment changes.
