+++ src/loaders/index.ts
@@ -1,8 +1,6 @@
-export { loadOmeTiff } from './tiff';
-export { loadTiffFolder } from './tiff-folder';
+export { loadOmeTiff, loadMultiTiff } from './tiff';
 export { loadBioformatsZarr, loadOmeZarr } from './zarr';
 export { getChannelStats } from './utils';
 
 export { default as TiffPixelSource } from './tiff/pixel-source';
 export { default as ZarrPixelSource } from './zarr/pixel-source';
-export { default as TiffFolderPixelSource } from './tiff-folder/pixel-source';
