--- src/loaders/tiff-folder/pixel-source.ts
@@ -1,100 +1,0 @@
-import type { GeoTIFFImage } from 'geotiff';
-import type { TypedArray } from 'zarr';
-import { getImageSize, isInterleaved, SIGNAL_ABORTED } from '../utils';
-
-import type Pool from '../tiff/lib/Pool';
-import type {
-  PixelSource,
-  PixelSourceMeta,
-  SupportedDtype,
-  Labels,
-  RasterSelection,
-  TileSelection,
-  PixelData
-} from '../../types';
-
-type ReadRastersOptions = NonNullable<
-  Parameters<GeoTIFFImage['readRasters']>[0]
->;
-
-class TiffFolderPixelSource<S extends string[]> implements PixelSource<S> {
-  private _channelImages: GeoTIFFImage[];
-
-  constructor(
-    channelImages: GeoTIFFImage[],
-    public dtype: SupportedDtype,
-    public tileSize: number,
-    public shape: number[],
-    public labels: Labels<S>,
-    public meta?: PixelSourceMeta,
-    public pool?: Pool
-  ) {
-    this._channelImages = channelImages;
-  }
-
-  async getRaster({ selection, signal }: RasterSelection<S>) {
-    // @ts-ignore
-    const image = this._channelImages[selection.c];
-    return this._readRasters(image, { signal });
-  }
-
-  async getTile({ x, y, selection, signal }: TileSelection<S>) {
-    const { height, width } = this._getTileExtent(x, y);
-    const x0 = x * this.tileSize;
-    const y0 = y * this.tileSize;
-    const window = [x0, y0, x0 + width, y0 + height];
-    // @ts-ignore
-    const image = this._channelImages[selection.c];
-    return this._readRasters(image, { window, width, height, signal });
-  }
-
-  private async _readRasters(image: GeoTIFFImage, props?: ReadRastersOptions) {
-    const interleave = isInterleaved(this.shape);
-    const raster = await image.readRasters({
-      interleave,
-      ...props,
-      pool: this.pool
-    });
-
-    if (props?.signal?.aborted) {
-      throw SIGNAL_ABORTED;
-    }
-
-    /*
-     * geotiff.js returns objects with different structure
-     * depending on `interleave`. It's weird, but this seems to work.
-     */
-    const data = (interleave ? raster : raster[0]) as TypedArray;
-    return {
-      data,
-      width: (raster as TypedArray & { width: number }).width,
-      height: (raster as TypedArray & { height: number }).height
-    } as PixelData;
-  }
-
-  /*
-   * Computes tile size given x, y coord.
-   */
-  private _getTileExtent(x: number, y: number) {
-    const { height: zoomLevelHeight, width: zoomLevelWidth } = getImageSize(
-      this
-    );
-    let height = this.tileSize;
-    let width = this.tileSize;
-    const maxXTileCoord = Math.floor(zoomLevelWidth / this.tileSize);
-    const maxYTileCoord = Math.floor(zoomLevelHeight / this.tileSize);
-    if (x === maxXTileCoord) {
-      width = zoomLevelWidth % this.tileSize;
-    }
-    if (y === maxYTileCoord) {
-      height = zoomLevelHeight % this.tileSize;
-    }
-    return { height, width };
-  }
-
-  onTileError(err: Error) {
-    console.error(err);
-  }
-}
-
-export default TiffFolderPixelSource;
