+++ src/loaders/tiff/index.ts
@@ -4,8 +4,10 @@ import type { GeoTIFF } from 'geotiff';
 import { createOffsetsProxy, checkProxies } from './lib/proxies';
 import LZWDecoder from './lib/lzw-decoder';
 import Pool from './lib/Pool';
+import { parseFilename } from './lib/utils';
 
-import { load } from './ome-tiff';
+import { load as loadOme } from './ome-tiff';
+import { load as loadMulti, MultiTiffChannel } from './multi-tiff';
 
 addDecoder(5, () => LZWDecoder);
 
@@ -21,7 +23,7 @@ interface OmeTiffOptions extends TiffOptions {
 }
 
 type UnwrapPromise<T> = T extends Promise<infer Inner> ? Inner : T;
-type MultiImage = UnwrapPromise<ReturnType<typeof load>>; // get return-type from `load`
+type MultiImage = UnwrapPromise<ReturnType<typeof loadOme>>; // get return-type from `load`
 
 /** @ignore */
 export async function loadOmeTiff(
@@ -86,6 +88,77 @@ export async function loadOmeTiff(
    */
   checkProxies(tiff);
 
-  const loaders = pool ? await load(tiff, new Pool()) : await load(tiff);
+  const loaders = pool ? await loadOme(tiff, new Pool()) : await loadOme(tiff);
   return images === 'all' ? loaders : loaders[0];
+}
+
+const DEFAULT_MULTI_IMAGE_NAME = 'MultiTiff';
+
+/**
+ * Opens a folder of multiple tiffs each containing one image (no stacks, timepoints, or pyramids).
+ * Loads each tiff as a channel.
+ * Expects one of two possible inputs:
+ *
+ * string - A comma separated list of URLs.
+ *
+ * (File & { path: string })[] - A list of file objects paired with their paths. In this case
+ * the parent folder name of the first image will be used as the image name and the filenames
+ * will be used to as the channel names.
+ *
+ * @param {string | File[]} source url or files with paths
+ * @param {{ fetchOptions: (undefined | RequestInit) }} options
+ * @return {Promise<{ data: TiffFolderPixelSource[], metadata: ImageMeta }>} data source and associated metadata.
+ */
+export async function loadMultiTiff(
+  source: string | (File & { path: string })[],
+  opts: OmeTiffOptions = {}
+) {
+  let imageName: string | undefined;
+  const { pool = true } = opts;
+  const channels: MultiTiffChannel[] = [];
+
+  if (typeof source === 'string') {
+    // Load from a comma separated list of URLs pointing to flat TIFFs
+
+    const filepaths = source.split(',');
+    if (filepaths[0]) {
+      const splitPath = filepaths[0].split('/');
+      // We only want to get the image name from the path if the TIFF is in a folder.
+      // If the TIFF url is == 2, then the first part is http and the second part is the filename.
+      if (splitPath.length > 2) imageName = splitPath[-2];
+    }
+
+    for (const filepath of filepaths) {
+      const parsedFilename = parseFilename(filepath);
+      const extension = parsedFilename.extension?.toLowerCase();
+      if (extension === 'tif' || extension === 'tiff') {
+        const channelName = parsedFilename.name;
+        if (channelName) {
+          const tiff = await fromUrl(filepath, { cacheSize: Infinity });
+          channels.push({ name: channelName, tiff });
+        }
+      }
+    }
+  } else {
+    // Load from a list of file objects.
+
+    // Try to get the imageName from the path, otherwise set to a default.
+    if (source[0]) imageName = source[0].path.split('/')[-2];
+
+    for (const file of source) {
+      const { name } = parseFilename(file.path);
+      if (name) {
+        const tiff = await fromBlob(file);
+        channels.push({ name, tiff });
+      }
+    }
+  }
+  if (channels.length > 0) {
+    if (!imageName) imageName = DEFAULT_MULTI_IMAGE_NAME;
+    const loader = pool
+      ? await loadMulti(imageName, channels, new Pool())
+      : await loadMulti(imageName, channels);
+    return loader;
+  }
+  throw new Error('Unable to load image from provided TiffFolder source.');
 }
