+++ tests/loaders/index.spec.js
@@ -2,7 +2,7 @@ import './utils.spec';
 
 import './ome-tiff.spec';
 import './tiff-lib.spec';
-import './tiff-folder.spec';
+import './multi-tiff.spec';
 
 import './bioformats-zarr.spec';
 import './zarr-lib.spec';
